+++ Alalba_VK/src/Alalba_VK/Assets/Mesh.h
@@ -25,7 +25,7 @@ namespace Alalba
     const vk::Buffer& GetVertexbuffer() const  { return *m_vertexBuffer;}
     const vk::Buffer& GetIndexbuffer() const { return *m_indexBuffer; }
     const vk::BLAS& GetBLAS() const { return *m_blas; }
-    const uint32_t GetIndexCount() const { return m_indexCount; }
+    const size_t GetIndexCount() const { return m_indexCount; }
     const uint32_t GetInstanceCount() const  { return 1; }
     ~Mesh() {};
 
@@ -36,7 +36,7 @@ namespace Alalba
     std::unique_ptr<vk::BLAS> m_blas;
 
     void LoadModel(const std::string& file, std::vector<Vertex>& vertices, std::vector<uint32_t>& indices);
-    uint32_t m_vertexCount;
-    uint32_t m_indexCount;
+    size_t m_vertexCount;
+    size_t m_indexCount;
   };
 }
